a1882583/mobile/requirements.txt
mediapipe>=0.10.11
numpy>=1.24.4
opencv_contrib_python>=4.5.3.56
opencv_python>=4.10.0.84
paho_mqtt>=2.1.0
